Fixes a pretty significant bug.
I meant to use `for..of` loop since `this.blocks` is an array, but was using `for..in`.  Switched to oldschool `for` loop since `for..of` isn't available in ES5.

Also cleaned up the mechanism by which `getBlockByNumber` evaluates its input.  In particular, it should only accept hex encoded strings or "latest", "pending" and "earliest".